move more code out of monolith
